+++ Modules/Intent.Modules.Eventing.MassTransit/release-notes.md
@@ -1,3 +1,7 @@
+### Version 7.0.5
+
+- Improvement: SQL Lite ambient transaction suppression.
+
 ### Version 7.0.4
 
 - Improvement: Updated NuGet package versions.